smh, putting replace api / redux stuff in openEvent
