+++ frontend/reducers/eventReducer.js
@@ -29,10 +29,6 @@ export default function(state = eventInitial, action) {
       };
     case GOT_ATTENDS:
       return { ...state, selectedAttends: action.payload };
-    case REPLACE_RSVPS:
-      return { ...state, selectedRsvps: LOAD.LOADING };
-    case REPLACED_RSVPS:
-      return { ...state, selectedRsvps: action.payload };
     case RESET_EVENT:
       return eventInitial;
     default:
